+++ testcontainers-advanced-imagebuilder-demo/src/main/java/software/xdev/Application.java
@@ -1,11 +1,15 @@
 package software.xdev;
 
+import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.List;
 
+import org.apache.commons.compress.archivers.tar.TarArchiveEntry;
 import org.slf4j.LoggerFactory;
 
 import software.xdev.testcontainers.imagebuilder.AdvancedImageFromDockerFile;
 import software.xdev.testcontainers.imagebuilder.compat.DockerfileCOPYParentsEmulator;
+import software.xdev.testcontainers.imagebuilder.transfer.fcm.FileLinesContentModifier;
 
 
 public final class Application
@@ -25,13 +29,46 @@ public final class Application
 				".md",
 				".cmd",
 				"/renovate.json5",
-				// We need to keep the pom.xml as maven can't resolve the modules otherwise
-				"testcontainers-advanced-imagebuilder/src/**",
-				"testcontainers-advanced-imagebuilder-demo/src/**"
+				"testcontainers-advanced-imagebuilder/**",
+				"testcontainers-advanced-imagebuilder-demo/**"
 			)
 			.withDockerFilePath(Paths.get("../testcontainers-advanced-imagebuilder-demo/Dockerfile"))
 			.withBaseDir(Paths.get("../"))
-			.withDockerFileLinesModifier(new DockerfileCOPYParentsEmulator());
+			.withDockerFileLinesModifier(new DockerfileCOPYParentsEmulator())
+			// Only copy the required maven modules and remove the not required ones
+			.withTransferArchiveTARCompressorCustomizer(c -> c.withContentModifier(
+				new FileLinesContentModifier()
+				{
+					@Override
+					public boolean shouldApply(
+						final Path sourcePath,
+						final String targetPath,
+						final TarArchiveEntry tarArchiveEntry)
+					{
+						return "pom.xml".equals(targetPath);
+					}
+					
+					@Override
+					public List<String> modify(
+						final List<String> lines,
+						final Path sourcePath,
+						final String targetPath,
+						final TarArchiveEntry tarArchiveEntry)
+					{
+						return lines.stream()
+							// Only keep the dummy-app submodule as this is only needed for building
+							.filter(s -> !(s.contains("<module>testcontainers-advanced-imagebuilder")
+								&& !s.contains("<module>testcontainers-advanced-imagebuilder-dummy-app")))
+							.toList();
+					}
+					
+					@Override
+					public boolean isIdentical(final List<String> original, final List<String> created)
+					{
+						return original.size() == created.size();
+					}
+				}
+			));
 		
 		final String imageName = builder.get();