+++ testcontainers-advanced-imagebuilder/src/main/java/software/xdev/testcontainers/imagebuilder/transfer/fcm/FileLinesContentModifier.java
@@ -13,17 +13,31 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package software.xdev.testcontainers.imagebuilder.transfer;
+package software.xdev.testcontainers.imagebuilder.transfer.fcm;
 
 import java.io.IOException;
-import java.io.InputStream;
+import java.nio.charset.StandardCharsets;
+import java.nio.file.Files;
 import java.nio.file.Path;
+import java.util.List;
 
 import org.apache.commons.compress.archivers.tar.TarArchiveEntry;
 
 
-@FunctionalInterface
-public interface FileContentModifier
+public interface FileLinesContentModifier extends AdvancedFileContentModifier<List<String>>
 {
-	InputStream apply(Path path, TarArchiveEntry tarArchiveEntry) throws IOException;
+	@Override
+	default List<String> readOriginal(
+		final Path sourcePath,
+		final String targetPath,
+		final TarArchiveEntry tarArchiveEntry) throws IOException
+	{
+		return Files.readAllLines(sourcePath);
+	}
+	
+	@Override
+	default byte[] createInputStreamData(final List<String> data)
+	{
+		return String.join("\n", data).getBytes(StandardCharsets.UTF_8);
+	}
 }